Fix misleading filename extension to match content (bash)
AFAIK, you should only name bash scripts ".sh" if they are backwards
compatible (i.e. also runnable) with sh.

This script is using bash features unavailable in sh without shelltype
or version checking, so would very probably fail or work incorrectly if
run with sh.

Therefore naming it ".sh" is dangerously misleading.
(The hashbang is ignored if passed as a parameter to sh.)